+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Fixed
+- Don't output items in `ArraySchema` if empty
+
 ## [0.10.0]
 ### Added
 - Add intersecting label equivalence relation